requirements.txt
asgiref==3.6.0
Django==4.2
django-multiupload==0.6.1
Pillow==9.5.0
pip==21.2.3
setuptools==57.4.0
sqlparse==0.4.4